flaskr/requirements.txt
alembic==1.12.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.0.1
psycopg2==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
PyJWT==1.7.1
